--- libs/dashboard/dashboard-ui/src/components/search.tsx
@@ -1,13 +1,0 @@
-import { Input } from '@home-improvement-hub/shadcn';
-
-export function Search() {
-  return (
-    <div>
-      <Input
-        type="search"
-        placeholder="Search..."
-        className="md:w-[100px] lg:w-[300px]"
-      />
-    </div>
-  );
-}
