+++ libs/dashboard/dashboard-ui/src/index.ts
@@ -1,8 +1,8 @@
 export * from './lib/dashboard-ui';
-export * from './components/user-nav';
-export * from './components/user-nav';
-export * from './components/main-nav';
-export * from './components/search';
+export * from '../../../shared/shared-ui/src/lib/components/user-nav';
+export * from '../../../shared/shared-ui/src/lib/components/user-nav';
+export * from '../../../shared/shared-ui/src/lib/components/main-nav';
+export * from '../../../shared/shared-ui/src/lib/components/search';
 export * from './components/recent-sales';
 export * from './components/overview';
 export * from './components/dashboard-nav';
