+++ libs/dashboard/features/feature-analytics/src/lib/feature-analytics.tsx
@@ -1,5 +1,13 @@
 export function FeatureAnalytics() {
-  return <h1>Welcome to FeatureAnalytics!</h1>;
+  return (
+    <main className="p-4 h-auto pt-20">
+      <div className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-96 mb-4 bg-yellow-500"></div>
+      <div className="grid grid-cols-2 gap-4">
+        <div className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-48 md:h-72 bg-teal-500"></div>
+        <div className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-48 md:h-72 bg-rose-500"></div>
+      </div>
+    </main>
+  );
 }
 
 export default FeatureAnalytics;
