--- libs/dashboard/features/feature-overview/src/lib/feature-overview.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import FeatureOverview from './feature-overview';
-
-describe('FeatureOverview', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<FeatureOverview />);
-    expect(baseElement).toBeTruthy();
-  });
-});
