+++ libs/dashboard/features/feature-overview/tsconfig.lib.json
@@ -4,6 +4,7 @@
     "outDir": "../../../../dist/out-tsc",
     "types": [
       "node",
+
       "@nx/react/typings/cssmodule.d.ts",
       "@nx/react/typings/image.d.ts"
     ]
