--- libs/dashboard/features/feature-recent-sales/README.md
@@ -1,7 +1,0 @@
-# feature-recent-sales
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test feature-recent-sales` to execute the unit tests via [Jest](https://jestjs.io).
