--- libs/dashboard/features/feature-recent-sales/eslint.config.js
@@ -1,12 +1,0 @@
-const nx = require('@nx/eslint-plugin');
-const baseConfig = require('../../../../eslint.config.js');
-
-module.exports = [
-  ...baseConfig,
-  ...nx.configs['flat/react'],
-  {
-    files: ['**/*.ts', '**/*.tsx', '**/*.js', '**/*.jsx'],
-    // Override or add rules here
-    rules: {},
-  },
-];
