--- libs/dashboard/features/feature-recent-sales/jest.config.ts
@@ -1,11 +1,0 @@
-export default {
-  displayName: 'feature-recent-sales',
-  preset: '../../../../jest.preset.js',
-  transform: {
-    '^(?!.*\\.(js|jsx|ts|tsx|css|json)$)': '@nx/react/plugins/jest',
-    '^.+\\.[tj]sx?$': ['babel-jest', { presets: ['@nx/react/babel'] }],
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory:
-    '../../../../coverage/libs/dashboard/features/feature-recent-sales',
-};
