--- libs/dashboard/features/feature-recent-sales/project.json
@@ -1,9 +1,0 @@
-{
-  "name": "feature-recent-sales",
-  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/dashboard/features/feature-recent-sales/src",
-  "projectType": "library",
-  "tags": ["scope: dashboard", "type:feature"],
-  "// targets": "to see all targets run: nx show project feature-recent-sales --web",
-  "targets": {}
-}
