--- libs/dashboard/features/feature-recent-sales/src/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/feature-recent-sales';
