--- libs/dashboard/features/feature-recent-sales/src/lib/feature-recent-sales.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import FeatureRecentSales from './feature-recent-sales';
-
-describe('FeatureRecentSales', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<FeatureRecentSales />);
-    expect(baseElement).toBeTruthy();
-  });
-});
