--- libs/dashboard/features/feature-recent-sales/src/lib/feature-recent-sales.tsx
@@ -1,5 +1,0 @@
-export function FeatureRecentSales() {
-  return <h1>Welcome to FeatureRecentSales!</h1>;
-}
-
-export default FeatureRecentSales;
