--- libs/dashboard/features/feature-reports/jest.config.ts
@@ -1,11 +1,0 @@
-export default {
-  displayName: 'feature-reports',
-  preset: '../../../../jest.preset.js',
-  transform: {
-    '^(?!.*\\.(js|jsx|ts|tsx|css|json)$)': '@nx/react/plugins/jest',
-    '^.+\\.[tj]sx?$': ['babel-jest', { presets: ['@nx/react/babel'] }],
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory:
-    '../../../../coverage/libs/dashboard/features/feature-reports',
-};
