+++ libs/dashboard/features/feature-reports/project.json
@@ -3,7 +3,7 @@
   "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
   "sourceRoot": "libs/dashboard/features/feature-reports/src",
   "projectType": "library",
-  "tags": ["scope: dashboard", "type:feature"],
+  "tags": [],
   "// targets": "to see all targets run: nx show project feature-reports --web",
   "targets": {}
 }
