--- libs/dashboard/features/feature-reports/src/lib/feature-reports.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import FeatureReports from './feature-reports';
-
-describe('FeatureReports', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<FeatureReports />);
-    expect(baseElement).toBeTruthy();
-  });
-});
