+++ libs/dashboard/features/feature-reports/src/lib/feature-reports.tsx
@@ -1,5 +1,23 @@
 export function FeatureReports() {
-  return <h1>Welcome to FeatureReports!</h1>;
+  return (
+    <main className="p-4 h-auto pt-20">
+      <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-4 mb-4">
+        <div className="border-2 border-dashed border-gray-300 rounded-lg dark:border-gray-600 h-32 md:h-64 bg-gray-500"></div>
+        <div className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-32 md:h-64 bg-amber-500"></div>
+        <div className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-32 md:h-64 bg-red-500"></div>
+        <div className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-32 md:h-64 bg-green-500"></div>
+      </div>
+      <div className="grid grid-cols-2 gap-4 mb-4">
+        <div className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-48 md:h-72 bg-sky-500"></div>
+        <div className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-48 md:h-72 bg-purple-500"></div>
+      </div>
+      <div className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-96 mb-4 bg-yellow-500"></div>
+      <div className="grid grid-cols-2 gap-4">
+        <div className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-48 md:h-72 bg-teal-500"></div>
+        <div className="border-2 border-dashed rounded-lg border-gray-300 dark:border-gray-600 h-48 md:h-72 bg-rose-500"></div>
+      </div>
+    </main>
+  );
 }
 
 export default FeatureReports;
