+++ libs/dashboard/features/feature-reports/tsconfig.json
@@ -11,9 +11,6 @@
   "references": [
     {
       "path": "./tsconfig.lib.json"
-    },
-    {
-      "path": "./tsconfig.spec.json"
     }
   ],
   "extends": "../../../../tsconfig.base.json"
