+++ libs/shadcn/src/index.ts
@@ -47,6 +47,6 @@ export * from './components/collapsible';
 export * from './components/command';
 export * from './components/use-is-mobile';
 export * from './components/sidebar';
-export * from './components/header';
+export * from '../../shared/shared-ui/src/lib/components/header';
 
 export * from './components/context-menu';
