+++ libs/shared/shared-ui/src/index.ts
@@ -1,8 +1,5 @@
 export * from './lib/ui';
-export * from './lib/layout-dashboard';
 export * from './lib/layout';
-export * from './lib/layout-orders';
 export * from './lib/data/schema';
-export * from './lib/components/user-nav';
 export * from './lib/components/data-table/columns';
 export * from './lib/components/data-table/data-table';
