+++ libs/shared/shared-ui/src/lib/components/header.tsx
@@ -1,11 +1,36 @@
-import { MainNav } from './main-nav';
+import { MainNav } from '@home-improvement-hub/dashboard-ui';
+import { Search } from '@home-improvement-hub/dashboard-ui';
+import { UserNav } from '@home-improvement-hub/dashboard-ui';
 
 export function Header() {
   return (
-    <div className="border-b">
-      <div className="flex h-16 items-center px-4">
-        <MainNav className="mx-6" />
-        <div className="ml-auto flex items-center space-x-4"></div>
+    <div>
+      <div className="md:hidden">
+        <img
+          src="/examples/dashboard-light.png"
+          width={1280}
+          height={866}
+          alt="Dashboard"
+          className="block dark:hidden"
+        />
+        <img
+          src="/examples/dashboard-dark.png"
+          width={1280}
+          height={866}
+          alt="Dashboard"
+          className="hidden dark:block"
+        />
+      </div>
+      <div className="hidden flex-col md:flex">
+        <div className="border-b">
+          <div className="flex h-16 items-center px-4">
+            <MainNav className="mx-6" />
+            <div className="ml-auto flex items-center space-x-4">
+              <Search />
+              <UserNav />
+            </div>
+          </div>
+        </div>
       </div>
     </div>
   );
