+++ libs/user-management/features/user-list/feature-edit-user/src/lib/feature-edit-user.tsx
@@ -1,15 +1,5 @@
-import styled from 'tailwind';
-
-const StyledFeatureEditUser = styled.div`
-  color: pink;
-`;
-
 export function FeatureEditUser() {
-  return (
-    <StyledFeatureEditUser>
-      <h1>Welcome to FeatureEditUser!</h1>
-    </StyledFeatureEditUser>
-  );
+  return <h1>Welcome to FeatureEditUser!</h1>;
 }
 
 export default FeatureEditUser;
