+++ libs/user-management/user-management-data-access/src/lib/user-management-data-access.tsx
@@ -1,15 +1,5 @@
-import styled from 'tailwind';
-
-const StyledUserManagementDataAccess = styled.div`
-  color: pink;
-`;
-
 export function UserManagementDataAccess() {
-  return (
-    <StyledUserManagementDataAccess>
-      <h1>Welcome to UserManagementDataAccess!</h1>
-    </StyledUserManagementDataAccess>
-  );
+  return <h1>Welcome to UserManagementDataAccess!</h1>;
 }
 
 export default UserManagementDataAccess;
