+++ libs/user-management/user-management-routing/src/lib/user-management-routing.tsx
@@ -1,15 +1,5 @@
-import styled from 'tailwind';
-
-const StyledUserManagementRouting = styled.div`
-  color: pink;
-`;
-
 export function UserManagementRouting() {
-  return (
-    <StyledUserManagementRouting>
-      <h1>Welcome to UserManagementRouting!</h1>
-    </StyledUserManagementRouting>
-  );
+  return <h1>Welcome to UserManagementRouting!</h1>;
 }
 
 export default UserManagementRouting;
